requirements.txt
mysql_connector_repackaged==0.3.1
pandas==2.2.3
python-dotenv==1.0.1
reportlab==4.2.5
streamlit==1.39.0
